requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
pandas>=1.3.0
numpy>=1.19.0
tqdm>=4.62.0
pickle-mixin>=1.0.2
matplotlib>=3.4.0
scikit-learn>=0.24.0